refactor beforeUnload to a hook
